refactor: move Pagination to separate folder and create corresponding css file
