+++ src/components/Home.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import Search from './Search/Search';
 import PokemonList from './PokemonList';
-import Pagination from '../components/Pagination';
+import Pagination from './Pagination/Pagination';
 import PokemonDetails from '../components/PokemonDetails';
 
 const Home: React.FC = () => {
