+++ src/components/Pagination/Pagination.css
@@ -1,0 +1,3 @@
+.pagination span {
+  padding: 0 5px;
+}
