+++ src/components/Pagination/Pagination.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { useNavigate, useLocation } from 'react-router-dom';
+import './Pagination.css';
 
 interface PaginationProps {
   currentPage: number;
